create new utils.cleverhans module to hold shared functionality for scripts using cleverhans stuff
